+++ core/src/main/kotlin/sift/core/graphviz/Dot.kt
@@ -1,0 +1,6 @@
+package sift.core.graphviz
+
+// needed for serialization...
+enum class Dot {
+    edge, node
+}