+++ sift.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env zsh
+#!/usr/bin/env bash
 
 # prereq: mvn clean install, alt mvn clean install -P native-image (graalvm)
 SIFT_PATH="$HOME/.local/share/sift"
@@ -6,30 +6,12 @@ SIFT_JAR="$SIFT_PATH/bin/sift-cli.jar"
 SIFT_BIN="$SIFT_PATH/bin/sift"
 
 # common args
-SIFT_ARGS=(
-    --ansi=truecolor
-)
-
-# base graphviz dot args
-DOT_ARGS=(-Tpng -Gmargin=0)
+SIFT_ARGS="--ansi=truecolor"
 
 # error codes
 ERROR_SIFT_NOT_FOUND=2
 ERROR_GRAPHVIZ_NOT_FOUND=3
 ERROR_IMAGE_VIEWERS_NOT_FOUND=4
-
-# set SIFT_DEBUG to any value to collect profiling data for use with
-# kcachegrind: building with the 'native-image-debug' maven profile is
-# a prerequisite for source code-level analysis
-if (( ${+SIFT_DEBUG} )) ; then
-    SIFT_VALGRIND=(
-        valgrind
-            --tool=callgrind
-            --dump-instr=yes
-            --simulate-cache=yes
-            --collect-jumps=yes
-    )
-fi
 
 function _sift() {
     if [[ -x $SIFT_BIN ]]; then
@@ -44,7 +26,7 @@ function _sift() {
 
 function pipe_to_dot() {
     # checks for --render in args
-    [[ ${*[(ie)--render]} -le ${#*} ]]
+    [[ "$@" =~ "--render" ]]
 }
 
 function exists() {
@@ -59,19 +41,16 @@ if pipe_to_dot $* ; then
     fi
 
     if [[ $TERM == "xterm-kitty" ]]; then
-        SHOW_IMG=(kitty +kitten icat --align left)
+        _sift $* | dot -Tpng -Gmargin=0 | kitty +kitten icat --align left
     elif exists feh ; then
-        SHOW_IMG=(feh --auto-zoom --scale-down -)
-        DOT_ARGS+=(-Gbgcolor=black)
+        _sift $* | dot -Tpng -Gmargin=0 -Gbgcolor=black | feh --auto-zoom --scale-down -
     elif exists display ; then
-        SHOW_IMG=(display)
-        DOT_ARGS+=(-Gbgcolor=black)
+        _sift $* | dot -Tpng -Gmargin=0 -Gbgcolor=black | display
     else
         echo "Error: Unable to find image viewers 'feh' or 'display'" >&2
         exit $ERROR_IMAGE_VIEWERS_NOT_FOUND
     fi
 
-    _sift $* | dot $DOT_ARGS | ${=SHOW_IMG}
 else
     _sift $* | less -FXRS
 fi